- Finished nav bar - Updated other files css reference
