+++ engage-concordia/src/Component/CSS/navbar.css
@@ -1,0 +1,71 @@
+.Navbar-container{
+    display: inline-flex;
+    justify-content: space-between;
+    width:100%;
+    height: 80px;
+    
+    background-color: #912338;
+
+    font-family: 'Inter', sans-serif;
+    font-size: 32px;
+font-style: normal;
+font-weight: 700;
+line-height: 140.625%; /* 45px */ 
+    color: #fff;
+}
+
+.app-name{
+    position: relative;
+    top: 28px; 
+    left: -77px;
+}
+
+button{
+    background-color: transparent;
+    background-repeat: no-repeat;
+    border: none;
+    cursor: pointer;
+    outline: none;
+    margin: 0;
+    padding:0;
+}
+
+.calendar-button-text {
+    position:relative;
+    padding: 0;
+    margin: 0;
+    left: 70px;
+    top: 8px; 
+}
+
+.profile-button-container{
+    display: inline-flex;
+}
+.profile-button-text{
+    
+    position: absolute;
+    top: 12px; 
+    right: 70px;
+    
+}
+
+.profile-button-text,.calendar-button-text{
+    font-family: 'Inter', sans-serif;
+    font-size: 26px;
+    font-style: normal;
+    font-weight: 700;
+    line-height: normal; 
+    color: #fff;
+}
+
+img{
+    position: absolute;
+    border-radius: 100%;
+    width: 56px;
+    height: 56px;
+    size: 56px;
+    top: 2px;
+    right: 7px;
+    
+}
+
