+++ engage-concordia/src/Component/Calendar.js
@@ -1,10 +1,11 @@
 import Navbar from './Navbar'
 
 const Calendar = () =>{
+  var styling =require('./CSS/calendar.css');
     return(<>
     <title>Your Calendar - Engage Concordia</title>
     <Navbar/>
-      <link rel="stylesheet" type="text/css" href="calendar.scss"></link>
+      <link rel="stylesheet" type="text/css" href={styling}></link>
       <h2>Calendar Page</h2>
     </>);
 };
